requirements.txt
pushbullet.py==0.12.0
selenium==3.141.0